requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
black==24.4.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
dj-database-url==2.2.0
Django==5.2
django-cors-headers==4.3.1
django-redis==5.4.0
django_urls==1.1.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.18.0
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
httpx-sse==0.4.0
identify==2.6.10
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mcp==1.7.1
mypy_extensions==1.1.0
nodeenv==1.9.1
openai==1.78.0
openai-agents==0.0.14
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
pre_commit==4.2.0
psycopg==3.2.7
psycopg-binary==3.2.7
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.2.0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
ruff==0.4.4
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==2.3.4
starlette==0.46.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
